+++ src/containers/SearchPage/SearchPageWithGrid.js
@@ -26,11 +26,10 @@ import { Footer, H2, Page } from '../../components';
 import TopbarContainer from '../TopbarContainer/TopbarContainer';
 
 import {
-  groupExtendedDataConfigs,
+  groupListingFieldConfigs,
   initialValues,
   searchParamsPicker,
   validUrlQueryParamsFromProps,
-  validURLParamsForExtendedData,
   validFilterParams,
   cleanSearchFromConflictingParams,
   createSearchResultSchema,
@@ -87,14 +86,11 @@ export class SearchPageComponent extends Component {
   // Reset all filter query parameters
   resetAll(e) {
     const { history, routeConfiguration, config } = this.props;
-    const { listingExtendedData: listingExtendedDataConfig } = config?.listing || {};
+    const { listingFields: listingFieldsConfig } = config?.listing || {};
     const { defaultFilters: defaultFiltersConfig } = config?.search || {};
 
     const urlQueryParams = validUrlQueryParamsFromProps(this.props);
-    const filterQueryParamNames = getQueryParamNames(
-      listingExtendedDataConfig,
-      defaultFiltersConfig
-    );
+    const filterQueryParamNames = getQueryParamNames(listingFieldsConfig, defaultFiltersConfig);
 
     // Reset state
     this.setState({ currentQueryParams: {} });
@@ -106,7 +102,7 @@ export class SearchPageComponent extends Component {
 
   getHandleChangedValueFn(useHistoryPush) {
     const { history, routeConfiguration, config } = this.props;
-    const { listingExtendedData: listingExtendedDataConfig } = config?.listing || {};
+    const { listingFields: listingFieldsConfig } = config?.listing || {};
     const { defaultFilters: defaultFiltersConfig, sortConfig } = config?.search || {};
 
     const urlQueryParams = validUrlQueryParamsFromProps(this.props);
@@ -137,7 +133,7 @@ export class SearchPageComponent extends Component {
           const searchParams = this.state.currentQueryParams;
           const search = cleanSearchFromConflictingParams(
             searchParams,
-            listingExtendedDataConfig,
+            listingFieldsConfig,
             defaultFiltersConfig,
             sortConfig
           );
@@ -185,7 +181,7 @@ export class SearchPageComponent extends Component {
       config,
     } = this.props;
 
-    const { listingExtendedData: listingExtendedDataConfig } = config?.listing || {};
+    const { listingFields: listingFieldsConfig } = config?.listing || {};
     const { defaultFilters: defaultFiltersConfig, sortConfig } = config?.search || {};
     const activeListingTypes = config?.listing?.listingTypes.map(config => config.listingType);
     const marketplaceCurrency = config.currency;
@@ -196,24 +192,25 @@ export class SearchPageComponent extends Component {
     const { searchParamsAreInSync, urlQueryParams, searchParamsInURL } = searchParamsPicker(
       location.search,
       searchParams,
-      listingExtendedDataConfig,
+      listingFieldsConfig,
       defaultFiltersConfig,
       sortConfig,
       isOriginInUse(config)
     );
 
-    const validQueryParams = validURLParamsForExtendedData(
+    const validQueryParams = validFilterParams(
       searchParamsInURL,
-      listingExtendedDataConfig,
-      defaultFiltersConfig
+      listingFieldsConfig,
+      defaultFiltersConfig,
+      false
     );
 
     const isKeywordSearch = isMainSearchTypeKeywords(config);
     const defaultFilters = isKeywordSearch
       ? defaultFiltersConfig.filter(f => f.key !== 'keywords')
       : defaultFiltersConfig;
-    const [customPrimaryFilters, customSecondaryFilters] = groupExtendedDataConfigs(
-      listingExtendedDataConfig,
+    const [customPrimaryFilters, customSecondaryFilters] = groupListingFieldConfigs(
+      listingFieldsConfig,
       activeListingTypes
     );
     const availableFilters = [
@@ -225,7 +222,7 @@ export class SearchPageComponent extends Component {
     // Selected aka active filters
     const selectedFilters = validFilterParams(
       validQueryParams,
-      listingExtendedDataConfig,
+      listingFieldsConfig,
       defaultFiltersConfig
     );
     const keysOfSelectedFilters = Object.keys(selectedFilters);
@@ -248,7 +245,7 @@ export class SearchPageComponent extends Component {
     const conflictingFilterActive = isAnyFilterActive(
       sortConfig.conflictingFilters,
       validQueryParams,
-      listingExtendedDataConfig,
+      listingFieldsConfig,
       defaultFiltersConfig
     );
     const sortBy = mode => {